代码走查

+++ src/main/java/ttjj/dto/RankBizDataDiff.java
@@ -317,39 +317,39 @@ public class RankBizDataDiff {
      */
     private Long f223;
 
-    private Double LAST_NET;
-    private Double NET_MIN_1;
-    private Double NET_MIN_7;
-    private Double NET_MIN_14;
-    private Double NET_MIN_30;
-    private Double NET_MIN_60;
-    private Double NET_MIN_90;
-    private Double NET_MIN_180;
-    private Double NET_MIN_360;
-    private Double NET_MAX_1;
-    private Double NET_MAX_7;
-    private Double NET_MAX_14;
-    private Double NET_MAX_30;
-    private Double NET_MAX_60;
-    private Double NET_MAX_90;
-    private Double NET_MAX_180;
-    private Double NET_MAX_360;
-    private Double NET_MIN_CLOS_1;
-    private Double NET_MIN_CLOS_7;
-    private Double NET_MIN_CLOS_14;
-    private Double NET_MIN_CLOS_30;
-    private Double NET_MIN_CLOS_60;
-    private Double NET_MIN_CLOS_90;
-    private Double NET_MIN_CLOS_180;
-    private Double NET_MIN_CLOS_360;
-    private Double NET_MAX_CLOS_1;
-    private Double NET_MAX_CLOS_7;
-    private Double NET_MAX_CLOS_14;
-    private Double NET_MAX_CLOS_30;
-    private Double NET_MAX_CLOS_60;
-    private Double NET_MAX_CLOS_90;
-    private Double NET_MAX_CLOS_180;
-    private Double NET_MAX_CLOS_360;
+    private BigDecimal LAST_NET;
+    private BigDecimal NET_MIN_1;
+    private BigDecimal NET_MIN_7;
+    private BigDecimal NET_MIN_14;
+    private BigDecimal NET_MIN_30;
+    private BigDecimal NET_MIN_60;
+    private BigDecimal NET_MIN_90;
+    private BigDecimal NET_MIN_180;
+    private BigDecimal NET_MIN_360;
+    private BigDecimal NET_MAX_1;
+    private BigDecimal NET_MAX_7;
+    private BigDecimal NET_MAX_14;
+    private BigDecimal NET_MAX_30;
+    private BigDecimal NET_MAX_60;
+    private BigDecimal NET_MAX_90;
+    private BigDecimal NET_MAX_180;
+    private BigDecimal NET_MAX_360;
+    private BigDecimal NET_MIN_CLOS_1;
+    private BigDecimal NET_MIN_CLOS_7;
+    private BigDecimal NET_MIN_CLOS_14;
+    private BigDecimal NET_MIN_CLOS_30;
+    private BigDecimal NET_MIN_CLOS_60;
+    private BigDecimal NET_MIN_CLOS_90;
+    private BigDecimal NET_MIN_CLOS_180;
+    private BigDecimal NET_MIN_CLOS_360;
+    private BigDecimal NET_MAX_CLOS_1;
+    private BigDecimal NET_MAX_CLOS_7;
+    private BigDecimal NET_MAX_CLOS_14;
+    private BigDecimal NET_MAX_CLOS_30;
+    private BigDecimal NET_MAX_CLOS_60;
+    private BigDecimal NET_MAX_CLOS_90;
+    private BigDecimal NET_MAX_CLOS_180;
+    private BigDecimal NET_MAX_CLOS_360;
 
     /**
      * NET_MA_5 净值-均线-n个
@@ -1512,267 +1512,267 @@ public class RankBizDataDiff {
         this.type = type;
     }
 
-    public Double getLAST_NET() {
+    public BigDecimal getLAST_NET() {
         return LAST_NET;
     }
 
-    public void setLAST_NET(Double LAST_NET) {
+    public void setLAST_NET(BigDecimal LAST_NET) {
         this.LAST_NET = LAST_NET;
     }
 
-    public Double getNET_MIN_1() {
+    public BigDecimal getNET_MIN_1() {
         return NET_MIN_1;
     }
 
-    public void setNET_MIN_1(Double NET_MIN_1) {
+    public void setNET_MIN_1(BigDecimal NET_MIN_1) {
         this.NET_MIN_1 = NET_MIN_1;
     }
 
-    public Double getNET_MIN_7() {
+    public BigDecimal getNET_MIN_7() {
         return NET_MIN_7;
     }
 
-    public void setNET_MIN_7(Double NET_MIN_7) {
+    public void setNET_MIN_7(BigDecimal NET_MIN_7) {
         this.NET_MIN_7 = NET_MIN_7;
     }
 
-    public Double getNET_MIN_14() {
+    public BigDecimal getNET_MIN_14() {
         return NET_MIN_14;
     }
 
-    public void setNET_MIN_14(Double NET_MIN_14) {
+    public void setNET_MIN_14(BigDecimal NET_MIN_14) {
         this.NET_MIN_14 = NET_MIN_14;
     }
 
-    public Double getNET_MIN_30() {
+    public BigDecimal getNET_MIN_30() {
         return NET_MIN_30;
     }
 
-    public void setNET_MIN_30(Double NET_MIN_30) {
+    public void setNET_MIN_30(BigDecimal NET_MIN_30) {
         this.NET_MIN_30 = NET_MIN_30;
     }
 
-    public Double getNET_MIN_60() {
+    public BigDecimal getNET_MIN_60() {
         return NET_MIN_60;
     }
 
-    public void setNET_MIN_60(Double NET_MIN_60) {
+    public void setNET_MIN_60(BigDecimal NET_MIN_60) {
         this.NET_MIN_60 = NET_MIN_60;
     }
 
-    public Double getNET_MIN_90() {
+    public BigDecimal getNET_MIN_90() {
         return NET_MIN_90;
     }
 
-    public void setNET_MIN_90(Double NET_MIN_90) {
+    public void setNET_MIN_90(BigDecimal NET_MIN_90) {
         this.NET_MIN_90 = NET_MIN_90;
     }
 
-    public Double getNET_MIN_180() {
+    public BigDecimal getNET_MIN_180() {
         return NET_MIN_180;
     }
 
-    public void setNET_MIN_180(Double NET_MIN_180) {
+    public void setNET_MIN_180(BigDecimal NET_MIN_180) {
         this.NET_MIN_180 = NET_MIN_180;
     }
 
-    public Double getNET_MIN_360() {
+    public BigDecimal getNET_MIN_360() {
         return NET_MIN_360;
     }
 
-    public void setNET_MIN_360(Double NET_MIN_360) {
+    public void setNET_MIN_360(BigDecimal NET_MIN_360) {
         this.NET_MIN_360 = NET_MIN_360;
     }
 
-    public Double getNET_MAX_1() {
+    public BigDecimal getNET_MAX_1() {
         return NET_MAX_1;
     }
 
-    public void setNET_MAX_1(Double NET_MAX_1) {
+    public void setNET_MAX_1(BigDecimal NET_MAX_1) {
         this.NET_MAX_1 = NET_MAX_1;
     }
 
-    public Double getNET_MAX_7() {
+    public BigDecimal getNET_MAX_7() {
         return NET_MAX_7;
     }
 
-    public void setNET_MAX_7(Double NET_MAX_7) {
+    public void setNET_MAX_7(BigDecimal NET_MAX_7) {
         this.NET_MAX_7 = NET_MAX_7;
     }
 
-    public Double getNET_MAX_14() {
+    public BigDecimal getNET_MAX_14() {
         return NET_MAX_14;
     }
 
-    public void setNET_MAX_14(Double NET_MAX_14) {
+    public void setNET_MAX_14(BigDecimal NET_MAX_14) {
         this.NET_MAX_14 = NET_MAX_14;
     }
 
-    public Double getNET_MAX_30() {
+    public BigDecimal getNET_MAX_30() {
         return NET_MAX_30;
     }
 
-    public void setNET_MAX_30(Double NET_MAX_30) {
+    public void setNET_MAX_30(BigDecimal NET_MAX_30) {
         this.NET_MAX_30 = NET_MAX_30;
     }
 
-    public Double getNET_MAX_60() {
+    public BigDecimal getNET_MAX_60() {
         return NET_MAX_60;
     }
 
-    public void setNET_MAX_60(Double NET_MAX_60) {
+    public void setNET_MAX_60(BigDecimal NET_MAX_60) {
         this.NET_MAX_60 = NET_MAX_60;
     }
 
-    public Double getNET_MAX_90() {
+    public BigDecimal getNET_MAX_90() {
         return NET_MAX_90;
     }
 
-    public void setNET_MAX_90(Double NET_MAX_90) {
+    public void setNET_MAX_90(BigDecimal NET_MAX_90) {
         this.NET_MAX_90 = NET_MAX_90;
     }
 
-    public Double getNET_MAX_180() {
+    public BigDecimal getNET_MAX_180() {
         return NET_MAX_180;
     }
 
-    public void setNET_MAX_180(Double NET_MAX_180) {
+    public void setNET_MAX_180(BigDecimal NET_MAX_180) {
         this.NET_MAX_180 = NET_MAX_180;
     }
 
-    public Double getNET_MAX_360() {
+    public BigDecimal getNET_MAX_360() {
         return NET_MAX_360;
     }
 
-    public void setNET_MAX_360(Double NET_MAX_360) {
+    public void setNET_MAX_360(BigDecimal NET_MAX_360) {
         this.NET_MAX_360 = NET_MAX_360;
     }
 
-    public Double getNET_MIN_CLOS_1() {
+    public BigDecimal getNET_MIN_CLOS_1() {
         return NET_MIN_CLOS_1;
     }
 
-    public void setNET_MIN_CLOS_1(Double NET_MIN_CLOS_1) {
+    public void setNET_MIN_CLOS_1(BigDecimal NET_MIN_CLOS_1) {
         this.NET_MIN_CLOS_1 = NET_MIN_CLOS_1;
     }
 
-    public Double getNET_MIN_CLOS_7() {
+    public BigDecimal getNET_MIN_CLOS_7() {
         return NET_MIN_CLOS_7;
     }
 
-    public void setNET_MIN_CLOS_7(Double NET_MIN_CLOS_7) {
+    public void setNET_MIN_CLOS_7(BigDecimal NET_MIN_CLOS_7) {
         this.NET_MIN_CLOS_7 = NET_MIN_CLOS_7;
     }
 
-    public Double getNET_MIN_CLOS_14() {
+    public BigDecimal getNET_MIN_CLOS_14() {
         return NET_MIN_CLOS_14;
     }
 
-    public void setNET_MIN_CLOS_14(Double NET_MIN_CLOS_14) {
+    public void setNET_MIN_CLOS_14(BigDecimal NET_MIN_CLOS_14) {
         this.NET_MIN_CLOS_14 = NET_MIN_CLOS_14;
     }
 
-    public Double getNET_MIN_CLOS_30() {
+    public BigDecimal getNET_MIN_CLOS_30() {
         return NET_MIN_CLOS_30;
     }
 
-    public void setNET_MIN_CLOS_30(Double NET_MIN_CLOS_30) {
+    public void setNET_MIN_CLOS_30(BigDecimal NET_MIN_CLOS_30) {
         this.NET_MIN_CLOS_30 = NET_MIN_CLOS_30;
     }
 
-    public Double getNET_MIN_CLOS_60() {
+    public BigDecimal getNET_MIN_CLOS_60() {
         return NET_MIN_CLOS_60;
     }
 
-    public void setNET_MIN_CLOS_60(Double NET_MIN_CLOS_60) {
+    public void setNET_MIN_CLOS_60(BigDecimal NET_MIN_CLOS_60) {
         this.NET_MIN_CLOS_60 = NET_MIN_CLOS_60;
     }
 
-    public Double getNET_MIN_CLOS_90() {
+    public BigDecimal getNET_MIN_CLOS_90() {
         return NET_MIN_CLOS_90;
     }
 
-    public void setNET_MIN_CLOS_90(Double NET_MIN_CLOS_90) {
+    public void setNET_MIN_CLOS_90(BigDecimal NET_MIN_CLOS_90) {
         this.NET_MIN_CLOS_90 = NET_MIN_CLOS_90;
     }
 
-    public Double getNET_MIN_CLOS_180() {
+    public BigDecimal getNET_MIN_CLOS_180() {
         return NET_MIN_CLOS_180;
     }
 
-    public void setNET_MIN_CLOS_180(Double NET_MIN_CLOS_180) {
+    public void setNET_MIN_CLOS_180(BigDecimal NET_MIN_CLOS_180) {
         this.NET_MIN_CLOS_180 = NET_MIN_CLOS_180;
     }
 
-    public Double getNET_MIN_CLOS_360() {
+    public BigDecimal getNET_MIN_CLOS_360() {
         return NET_MIN_CLOS_360;
     }
 
-    public void setNET_MIN_CLOS_360(Double NET_MIN_CLOS_360) {
+    public void setNET_MIN_CLOS_360(BigDecimal NET_MIN_CLOS_360) {
         this.NET_MIN_CLOS_360 = NET_MIN_CLOS_360;
     }
 
-    public Double getNET_MAX_CLOS_1() {
+    public BigDecimal getNET_MAX_CLOS_1() {
         return NET_MAX_CLOS_1;
     }
 
-    public void setNET_MAX_CLOS_1(Double NET_MAX_CLOS_1) {
+    public void setNET_MAX_CLOS_1(BigDecimal NET_MAX_CLOS_1) {
         this.NET_MAX_CLOS_1 = NET_MAX_CLOS_1;
     }
 
-    public Double getNET_MAX_CLOS_7() {
+    public BigDecimal getNET_MAX_CLOS_7() {
         return NET_MAX_CLOS_7;
     }
 
-    public void setNET_MAX_CLOS_7(Double NET_MAX_CLOS_7) {
+    public void setNET_MAX_CLOS_7(BigDecimal NET_MAX_CLOS_7) {
         this.NET_MAX_CLOS_7 = NET_MAX_CLOS_7;
     }
 
-    public Double getNET_MAX_CLOS_14() {
+    public BigDecimal getNET_MAX_CLOS_14() {
         return NET_MAX_CLOS_14;
     }
 
-    public void setNET_MAX_CLOS_14(Double NET_MAX_CLOS_14) {
+    public void setNET_MAX_CLOS_14(BigDecimal NET_MAX_CLOS_14) {
         this.NET_MAX_CLOS_14 = NET_MAX_CLOS_14;
     }
 
-    public Double getNET_MAX_CLOS_30() {
+    public BigDecimal getNET_MAX_CLOS_30() {
         return NET_MAX_CLOS_30;
     }
 
-    public void setNET_MAX_CLOS_30(Double NET_MAX_CLOS_30) {
+    public void setNET_MAX_CLOS_30(BigDecimal NET_MAX_CLOS_30) {
         this.NET_MAX_CLOS_30 = NET_MAX_CLOS_30;
     }
 
-    public Double getNET_MAX_CLOS_60() {
+    public BigDecimal getNET_MAX_CLOS_60() {
         return NET_MAX_CLOS_60;
     }
 
-    public void setNET_MAX_CLOS_60(Double NET_MAX_CLOS_60) {
+    public void setNET_MAX_CLOS_60(BigDecimal NET_MAX_CLOS_60) {
         this.NET_MAX_CLOS_60 = NET_MAX_CLOS_60;
     }
 
-    public Double getNET_MAX_CLOS_90() {
+    public BigDecimal getNET_MAX_CLOS_90() {
         return NET_MAX_CLOS_90;
     }
 
-    public void setNET_MAX_CLOS_90(Double NET_MAX_CLOS_90) {
+    public void setNET_MAX_CLOS_90(BigDecimal NET_MAX_CLOS_90) {
         this.NET_MAX_CLOS_90 = NET_MAX_CLOS_90;
     }
 
-    public Double getNET_MAX_CLOS_180() {
+    public BigDecimal getNET_MAX_CLOS_180() {
         return NET_MAX_CLOS_180;
     }
 
-    public void setNET_MAX_CLOS_180(Double NET_MAX_CLOS_180) {
+    public void setNET_MAX_CLOS_180(BigDecimal NET_MAX_CLOS_180) {
         this.NET_MAX_CLOS_180 = NET_MAX_CLOS_180;
     }
 
-    public Double getNET_MAX_CLOS_360() {
+    public BigDecimal getNET_MAX_CLOS_360() {
         return NET_MAX_CLOS_360;
     }
 
-    public void setNET_MAX_CLOS_360(Double NET_MAX_CLOS_360) {
+    public void setNET_MAX_CLOS_360(BigDecimal NET_MAX_CLOS_360) {
         this.NET_MAX_CLOS_360 = NET_MAX_CLOS_360;
     }
